Only add window.Repo etc. in the browser (#115)
In f2afa44e we added some code which populated the following global
variables on every page:

* `window.Automerge` - the `@automerge/automerge` package
* `window.AutomergeRepo` - the `@automerge/automerge-repo` package
* `window.Repo` - an instance of `Repo` pointing at the default sync
  server

This makes it possible to mess around with automerge in the developer
console whilst reading the docs.

Unfortunately this does not play nice with Docusaurus' static site
generation. The problem is that the server side code generates an import
for the automerge package which doesn't work because it attempts to
import the webassembly from the filesystem. The solution is to only do
the imports in the client side package by checking
`ExecutionEnvironment.canUseDOM`.

This required updating Docusaurus as the import of
`ExecutionEnvironment` triggered some warnings.